requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
gunicorn==21.2.0
Flask-WTF==1.1.1
email_validator==2.0.0
werkzeug==2.3.7
Flask-Migrate==4.0.7
pandas==2.2.2
openpyxl==3.1.5
python-dotenv==1.0.1
numpy==1.26.4
pdfkit==1.0.0
